requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
chromadb
openai
python-docx
jinja2
langchain
langchain-community
langchain-openai
sentence-transformers
pymupdf
ollama
mcp
pydantic-settings
python-multipart
reportlab
python-dotenv
